% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Minimum & Maximum method.R
\name{pdc_min_max_calc}
\alias{pdc_min_max_calc}
\title{A PDC calculation Function}
\usage{
pdc_min_max_calc(
  ID = "MRN",
  Data.Date = Sys.Date(),
  data_order = data_order,
  Order.Date = "Order.Date",
  Start.Date = "Start.Date",
  End.Date = "End.Date",
  data_dispensed = data_dispensed,
  Drug = "Drug.Name",
  Fill.Date = "Fill.Date",
  Days.Supplied = "Days.Supplied",
  data_visit = data_visit,
  Visit.Date = "Visit.Date",
  Death.Date = "Death.Date",
  Followed.Days = 180,
  data_hosp = data_hosp,
  Hosp.Start = "Hosp.Start",
  Hosp.End = "Hosp.End",
  parallel = T,
  ncores = 4
)
}
\arguments{
\item{ID}{Patients' ID}
}
\description{
This function allows you to calculate PDC in multiple measurements.
}
\examples{
pdc_min_max_calc()
}
\keyword{PDC}
